% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGCpct.R
\name{getGCpct}
\alias{getGCpct}
\title{Get GC percentages at specific genomic positions}
\usage{
getGCpct(chrom, pos, gbuild = c("hg19", "hg38"))
}
\arguments{
\item{chrom}{Chromosome of interest. Only one chromosome can be
specified at a time. For humans it can be 1-22, X(23), Y(24) or MT(25).}

\item{pos}{Genomic positions in the chromosome of interest}

\item{gbuild}{Genome build. Available builds are hg19 and hg38 for humans.}
}
\value{
A vector of GC percentage values corresponding to the genomic positions.
}
\description{
Gets GC percentages at specific genomic positions
Computes the GC percentages for a set of genomic positions along a
chromosome. The GC percentage is the proportion of GC nucleotides in a
1000 base window centered at the genomic position.
}
\examples{
\dontrun{
pos <- sort(sample(249250000, 10))
getGCpct(1, pos)
getGCpct(1, pos, "hg19")
}
}
