% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19}
\alias{hg19}
\title{hg19 GC percentages}
\format{A list containing 25 vectors corresponding to chromosomes 1-22, X, Y &
MT in that order.}
\usage{
hg19
}
\description{
GC percentages in 1 kiolobase sliding windows along chromosomes 1-22,
X, Y & MT for human genome build hg19.
}
\keyword{datasets}
