% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebbles-package.R
\docType{package}
\name{pebbles-package}
\alias{pebbles-package}
\title{pebbles: test data for rock}
\description{
Contains cancer genomics datasets for running tests and examples in the
rock R package (https://github.com/pdiakumis/rock)
}
